feat(resolve_n_vist): Resolver and Interpreter mostly done.
The code changes that were made are likely extremely broken. The biggest
issue here is the fact that I didn't test the codebase. There's a few
approaches to this problem. The largest way to solve this problem is by
running the code through a few ML & Symbolic analysis tools. I can do a
few on GitHub.

BREAKING CHANGE: UNKNOWN